% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_miss_forest.R
\name{add_imputed_to_og}
\alias{add_imputed_to_og}
\title{Add Imputed Data to Original Datasets}
\usage{
add_imputed_to_og(og_dfs, best_imp_outputs)
}
\arguments{
\item{og_dfs}{List of original surveys before imputation}

\item{best_imp_outputs}{List of best outputs from missForest algorithm}
}
\value{
List of new, imputed surveys with readout of NAs before and after
}
\description{
Helper function to put the imputed data back into the original data.
Removes demographics from imputed datasets (we don't want imputed demographics),
then combines with original survey data.
}
\examples{
\dontrun{
# Example usage
final_data <- add_imputed_to_og(original_surveys, best_imputations)
}
}
