% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_plot.R
\name{ba_plot}
\alias{ba_plot}
\title{Bland-Altman Plot}
\usage{
ba_plot(m1, m2)
}
\arguments{
\item{m1}{First set of measurements}

\item{m2}{Second set of measurements}
}
\value{
Creates a Bland-Altman plot showing agreement between two measurement methods
}
\description{
Slightly modified function from Karun and Puranik 2021. Mostly just
changed some formatting - margins, text locations, axis limits.
Reference: https://cegh.net/article/S2213-3984(21)00139-1/fulltext
}
\examples{
\dontrun{
# Example usage with two measurement methods
method1 <- rnorm(50, mean = 10, sd = 2)
method2 <- method1 + rnorm(50, mean = 0, sd = 0.5)
ba_plot(method1, method2)
}
}
