% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_miss_forest.R
\name{choose_best_runs}
\alias{choose_best_runs}
\title{Choose Best MissForest Runs}
\usage{
choose_best_runs(imp_surveys, tuning_grid = full_grid, variablewise = TRUE)
}
\arguments{
\item{imp_surveys}{List of imputation results from run_miss_forest}

\item{tuning_grid}{Data frame with hyperparameter combinations used}

\item{variablewise}{Logical, whether variable-wise errors were computed (default: TRUE)}
}
\value{
List of best imputation results with lowest error rates
}
\description{
Helper function to pull the best runs out of the list of all the imputations
based on lowest error rates.
}
\examples{
\dontrun{
# Example usage  
best_results <- choose_best_runs(imp_results, tuning_grid)
}
}
