% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_data.R
\name{clear_data}
\alias{clear_data}
\title{Clear Values from Global Environment}
\usage{
clear_data()
}
\description{
Function to remove values from the global environment,
useful for cleaning up after scripts while preserving loaded functions.
}
\examples{
\dontrun{
# Create some test objects
x <- 1:10
y <- "test"

# Remove data objects but keep functions
clear_data()
}
}
