% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_cond_fit.R
\name{conditional-fit-functions}
\alias{conditional-fit-functions}
\title{Iterative Conditional Fit Functions}
\description{
Functions for running iterative conditional fit statistics with bootstrapped
fit statistics from iarm::boot_fit, removing items with significant fit stats,
and rerunning the CML Rasch model with reduced items.
}
