% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cond_rebl_items.R
\name{get_cond_rebl_items}
\alias{get_cond_rebl_items}
\title{Get Conditional REBL Items}
\usage{
get_cond_rebl_items(survey_2a, survey_2b)
}
\arguments{
\item{survey_2a}{Conditional fit output from \code{boot_fit()} for survey 2a}

\item{survey_2b}{Conditional fit output from \code{boot_fit()} for survey 2b}
}
\value{
A dataframe that contains:
\itemize{
\item REBL items that were shared between surveys 2a and 2b
\item p values for those items from survey 2a
\item p values for those items from survey 2b
}
}
\description{
Function to explore iterative conditional fit results, see which items are
shared between surveys 2a and 2b, and check their p values.
}
\examples{
\dontrun{
# Example usage
shared_items <- get_cond_rebl_items(fit_2a, fit_2b)
}
}
