% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fits.R, R/rasch_test_functions.R
\name{get_fits}
\alias{get_fits}
\title{Get person and item fits}
\usage{
get_fits(model)

get_fits(model)
}
\arguments{
\item{model}{A Rasch model object from eRm}
}
\value{
A list containing person fits and item fits.

List containing person_fit and item_fit statistics
}
\description{
Using an eRm model, get person parameter fit and item fit
objects.

Get both item and person fit statistics from Rasch model.
}
\examples{
\dontrun{
  model <- eRm::RM(df)
  fits <- get_fits(model)
}
\dontrun{
# Get fit statistics
fits <- get_fits(rasch_model)
}
}
\seealso{
\code{\link[=test_lr_gender]{test_lr_gender()}}, \code{\link[=test_lr_rurality]{test_lr_rurality()}}, \code{\link[=test_lr_race]{test_lr_race()}}
}
