% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_test.R
\name{get_lr_test}
\alias{get_lr_test}
\title{Get Likelihood Ratio Test}
\usage{
get_lr_test(survey, var, rebl_names)
}
\arguments{
\item{survey}{Survey data frame}

\item{var}{A variable to use to filter for NAs and as split criterion (unquoted variable name)}

\item{rebl_names}{Vector of REBL item names. This can be all items, or a subset}
}
\value{
An LR test output from the eRm package
}
\description{
Function that gets a Likelihood Ratio test based on a split criterion. First, it filters
to make sure that variable has no NAs (necessary for LRTest), then it runs
a new \code{RM()}, then runs \code{LRTest()} with the split criterion given.
}
\examples{
\dontrun{
# Example usage
lr_result <- get_lr_test(survey_data, gender, rebl_item_names)
}
}
