% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_reduction_helpers.R
\name{get_rm}
\alias{get_rm}
\title{Get Rasch Models}
\usage{
get_rm(rebl_items, clean_surveys)
}
\arguments{
\item{rebl_items}{Vector of REBL item names}

\item{clean_surveys}{List of clean survey datasets}
}
\value{
List of Rasch models, one for each survey
}
\description{
Create Rasch models for specified REBL items across clean surveys.
}
\examples{
\dontrun{
# Example usage
models <- get_rm(final_items, clean_surveys)
}
}
