% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_size.R
\name{get_size}
\alias{get_size}
\title{Get Object Size}
\usage{
get_size(x)
}
\arguments{
\item{x}{An R object to get the size of}
}
\value{
A formatted string showing the object size in appropriate units
}
\description{
Wrapper around format(object.size(x)) to get it in Kb, Mb or Gb depending on size.
}
\examples{
\dontrun{
# Example usage
get_size(mtcars)
get_size(1:1000000)
}
}
