% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get Table with NAs}
\usage{
get_table(x)
}
\arguments{
\item{x}{Vector or factor to create table from}
}
\value{
Table object showing frequencies including NA values
}
\description{
Wrapper for table() that includes NA values by default (useNA = 'always').
}
\examples{
\dontrun{
# Example usage
get_table(c(1, 2, 1, NA, 2, 1))
}
}
