% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_reduction_helpers.R
\name{get_uni_test_multiple}
\alias{get_uni_test_multiple}
\title{Get Unidimensionality Test (Multiple Surveys)}
\usage{
get_uni_test_multiple(rebl_items, clean_surveys)
}
\arguments{
\item{rebl_items}{Vector of REBL item names}

\item{clean_surveys}{List of clean survey datasets}
}
\value{
List of NPtest results, one for each survey
}
\description{
Run unidimensionality tests for REBL items across multiple surveys.
}
\examples{
\dontrun{
# Example usage for multiple surveys
uni_tests <- get_uni_test_multiple(rebl_items, survey_list)
}
}
