% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_rebl_items.R
\name{id_rebl_items}
\alias{id_rebl_items}
\title{Identify REBL Items}
\usage{
id_rebl_items(df, pattern, ...)
}
\arguments{
\item{df}{A data frame containing all the REBL items}

\item{pattern}{A regex matching pattern}

\item{...}{Additional arguments passed to \code{grepl()}}
}
\value{
A character vector of the names of your REBL items.
}
\description{
Create a vector of your REBL items from a data frame using a
regex matching pattern.
}
\details{
This is a convenience function, but not essential in the workflow. If you
already have a vector of your REBL item names, that works just as well. Using
this function might be easier or harder depending on how items are named. If
they are something like \verb{rebl_*} this would work quite well. Otherwise, like
in the example, we might have to use some "OR" operators to represent them
all.

A vector of your REBL items will be required later in the workflow, however
you choose to get it.
}
\examples{
\dontrun{
data(example)
(id_rebl_items(example, 'lunch')
(id_rebl_items(example, 'lunch', ignore.case = TRUE))
(id_rebl_items(example, '^food')
(id_rebl_items(example, '^food|^pack|^home|^water|^social')
}
}
\seealso{
\code{\link[=id_reversed_rebl_items]{id_reversed_rebl_items()}}, \code{\link[=reverse_code_rebl_items]{reverse_code_rebl_items()}},
\code{\link[=get_rasch_model]{get_rasch_model()}}, \code{\link[=get_rebl_scores]{get_rebl_scores()}}
}
