% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_reversed_rebl_items.R
\name{id_reversed_rebl_items}
\alias{id_reversed_rebl_items}
\title{Identify Reversed REBL Items}
\usage{
id_reversed_rebl_items(rebl_items, pattern, ...)
}
\arguments{
\item{rebl_items}{A character vector of REBL item names}

\item{pattern}{A regex matching pattern to identify items that need reverse coding}

\item{...}{Additional arguments passed to \code{grepl()}}
}
\value{
A character vector of REBL item names that need reverse coding.
}
\description{
Filter a vector of REBL items to identify only those that need
reverse coding using a regex matching pattern.
}
\details{
This is a convenience function for identifying which REBL items need reverse
coding based on a pattern. For example, if items that need reverse coding
have "reverse" or "neg" in their names, you can use a pattern to identify them.
The function filters the provided vector of REBL items to return only those
matching the pattern.
}
\examples{
\dontrun{
  # First get all REBL items
  rebl_items <- id_rebl_items(example, '^rebl_')
  
  # Then identify which ones need reverse coding
  reversed_items <- id_reversed_rebl_items(rebl_items, 'reverse')
  reversed_items <- id_reversed_rebl_items(rebl_items, '_neg$')
  reversed_items <- id_reversed_rebl_items(rebl_items, 'item[135]$')
}
}
\seealso{
\code{\link[=id_rebl_items]{id_rebl_items()}}, \code{\link[=reverse_code_rebl_items]{reverse_code_rebl_items()}},
\code{\link[=get_rasch_model]{get_rasch_model()}}, \code{\link[=get_rebl_scores]{get_rebl_scores()}}
}
