% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_erm.R
\name{read_erm}
\alias{read_erm}
\title{Read eRm Objects}
\usage{
read_erm(x, loc.out = FALSE, as.irt.pars = TRUE)
}
\arguments{
\item{x}{A Rasch model object from eRm}

\item{loc.out}{Logical, location output parameter (default: FALSE)}

\item{as.irt.pars}{Logical, whether to return as IRT parameters (default: TRUE)}
}
\value{
IRT parameters object or parameter list
}
\description{
Modified version of plink::read.erm that actually works. Sets the groups
and time variables to 1 within the function, as these are not part of the
standard output from eRm objects.
}
\examples{
\dontrun{
# Convert eRm object to IRT parameters
irt_pars <- read_erm(rasch_model)
}
}
