% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_attention_checks.R
\name{recode_attention_checks}
\alias{recode_attention_checks}
\title{Recode Attention Checks}
\usage{
recode_attention_checks(df)
}
\arguments{
\item{df}{Survey dataframe with attention checks uncoded}
}
\value{
Survey dataframe with attention checks recoded as 1 for pass and 0 for fail
}
\description{
Function to recode attention checks. Note that Survey 1 only has 2, while
Surveys 2a and 2b have 4. The first two checks in each survey pass with NA,
while the second two checks only (Survey 2a and 2b) pass with 'Agree'.
}
\examples{
\dontrun{
# Example usage
recoded_data <- recode_attention_checks(survey_data)
}
}
