% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_demos.R
\name{recode_demos}
\alias{recode_demos}
\title{Recode Demographics}
\usage{
recode_demos(df)
}
\arguments{
\item{df}{Survey dataframe}
}
\value{
Dataframe with demographics recoded
}
\description{
Recode demographic questions for surveys 2a and 2b. Not sure the questions
are the same on surveys 1 and ewe validation set... will have to explore that.
}
\details{
Leaving rurality, election, and politics as categorical for now.
Age and income are currently midpoints
Education is binary (0 = no bachelor's degree, 1 = bachelor's degree or higher)
Gender is binary (0 = male, 1 = female/other)
Race is handled differently for survey 1 vs others
}
\examples{
\dontrun{
# Example usage
recoded_data <- recode_demos(survey_data)
}
}
