% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_likert.R
\name{recode_likert}
\alias{recode_likert}
\title{Recode Likert Scale Items}
\usage{
recode_likert(df)
}
\arguments{
\item{df}{Dataframe that contains any of these categories of Likert questions}
}
\value{
Same dataframe but recoded numeric 0 to 6 or 0 to 1 for binary
}
\description{
Automatically look for column names containing efficacy, ccBelief, enviroID,
greenGlow, worry, attribution, socialNorms, eweExpGen, psychDist, or hemp and
recode them as 7pt or binary as appropriate.
}
\details{
Note that this really only applies to surveys 2a and 2b. Different procedures
needed for survey 1 and ewe validation set.
}
\examples{
\dontrun{
# Example usage
recoded_data <- recode_likert(survey_data)
}
}
