% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_rebl.R
\name{recode_rebl}
\alias{recode_rebl}
\title{Recode REBL Items}
\usage{
recode_rebl(df, rebl_items)
}
\arguments{
\item{df}{A dataframe that includes all the REBL items}

\item{rebl_items}{A character vector of all REBL items. You can use
\code{\link[=id_rebl_items]{id_rebl_items()}} to get this.}
}
\value{
Same dataframe with "Yes" and "No" coded to 1 and 0 as factors based
on which one is the PEB
}
\description{
Recode REBL items from Yes/No to 1/0. Does not deal with reverse
coding. See \code{\link[=reverse_code_rebl_items]{reverse_code_rebl_items()}} for that.
}
\details{
Will show a warning if any of the REBL items provided do not appear
in the data frame.
}
\examples{
\dontrun{
  rebl_items <- id_rebl_items(raw_example, '^(?!res).*')
  df <- recode_rebl(raw_example, rebl_items)
}
}
\seealso{
\code{\link[=id_rebl_items]{id_rebl_items()}}, \code{\link[=id_reversed_rebl_items]{id_reversed_rebl_items()}},
\code{\link[=reverse_code_rebl_items]{reverse_code_rebl_items()}}, \code{\link[=get_rasch_model]{get_rasch_model()}}
}
