% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_code_rebl_items.R
\name{reverse_code_rebl_items}
\alias{reverse_code_rebl_items}
\title{Reverse Code REBL Items}
\usage{
reverse_code_rebl_items(df, reversed_items)
}
\arguments{
\item{df}{A dataframe containing the REBL items to be reverse coded}

\item{reversed_items}{A character vector of REBL item names that need to be
reverse coded}
}
\value{
A dataframe with the specified items reverse coded (1s become 0s,
0s become 1s, NAs remain NA)
}
\description{
Reverse code specified REBL items by flipping 1s to 0s and 0s to 1s.
This is typically done for items where "No" represents the pro-environmental
behavior rather than "Yes".
}
\examples{
\dontrun{
  # Identify items that need reverse coding
  reversed_items <- id_reversed_rebl_items(rebl_items, "pattern")
  df_reversed <- reverse_code_rebl_items(df, reversed_items)
}
}
\seealso{
\code{\link[=id_rebl_items]{id_rebl_items()}}, \code{\link[=id_reversed_rebl_items]{id_reversed_rebl_items()}},
\code{\link[=recode_rebl]{recode_rebl()}}, \code{\link[=get_rasch_model]{get_rasch_model()}}
}
