% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_test_functions.R
\name{test_local_dependence}
\alias{test_local_dependence}
\title{Test Local Dependence}
\usage{
test_local_dependence(survey, rebl_items)
}
\arguments{
\item{survey}{Survey dataset}

\item{rebl_items}{Vector of REBL item names}
}
\value{
List containing Q3 test results and proportion of coefficients above |0.3|
}
\description{
Test for local dependence using Q3 statistic.
}
\examples{
\dontrun{
# Test local dependence (this can be slow)
local_dep <- test_local_dependence(survey_data, rebl_items)
}
}
