% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_test_functions.R, R/test_lr_gender.R
\name{test_lr_gender}
\alias{test_lr_gender}
\title{Test LR by Gender}
\usage{
test_lr_gender(survey, rebl_names, var = gender)

test_lr_gender(survey, rebl_names, var = gender)
}
\arguments{
\item{survey}{Dataframe of containing REBL item responses}

\item{rebl_names}{A character vector of the names of the REBL items to be used
to filter the df to only REBL items.}

\item{var}{Currently only works for gender. Might add to this.}
}
\value{
LR test results

Test output from eRm::LRtest()
}
\description{
Test likelihood ratio by gender grouping.

Likelihood ratio test for invariance of eRm model splitting
sample by gender. NOTE: can we have this function run splits by different
variables?
}
\examples{
\dontrun{
# Test LR by gender
lr_gender <- test_lr_gender(survey_data, rebl_items)
}
}
