% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_test_functions.R
\name{test_lr_income}
\alias{test_lr_income}
\title{Test LR by Income}
\usage{
test_lr_income(survey, rebl_names, var = income, dummy_name = income_dummy)
}
\arguments{
\item{survey}{Survey dataset}

\item{rebl_names}{Vector of REBL item names}

\item{var}{Variable to use for grouping (default: income, unquoted)}

\item{dummy_name}{Name for dummy variable (default: income_dummy, unquoted)}
}
\value{
LR test results with income split summary
}
\description{
Test likelihood ratio by income level (split at $75k).
}
\examples{
\dontrun{
# Test LR by income
lr_income <- test_lr_income(survey_data, rebl_items)
}
}
