% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_test_functions.R
\name{test_rasch_model}
\alias{test_rasch_model}
\title{Test Rasch Model Comprehensively}
\usage{
test_rasch_model(surveys, rebl_items, models)
}
\arguments{
\item{surveys}{List of survey datasets}

\item{rebl_items}{Vector of REBL item names}

\item{models}{List of Rasch models (optional, will be created if NULL)}
}
\value{
List of comprehensive test results for each survey
}
\description{
Run all Rasch model tests together (except local dependence which is slow).
Tests unidimensionality, goodness of fit, invariance, separation reliability,
and demographic differences.
}
\examples{
\dontrun{
# Comprehensive model testing
test_results <- test_rasch_model(survey_list, rebl_items)
}
}
