% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_test_functions.R, R/test_uni_pcar.R
\name{test_uni_pcar}
\alias{test_uni_pcar}
\title{Test Unidimensionality with PCA of Residuals}
\usage{
test_uni_pcar(model, rotate = "Promax", n_factors = 5)

test_uni_pcar(model, rotate = "Promax", n_factors = 5)
}
\arguments{
\item{model}{An eRm model object.}

\item{rotate}{Name of rotation to use for PCAR. Passed to \code{pca()}.}

\item{n_factors}{Number of factors to extract}

\item{n_vars}{Number of variables (default: length of model$se.beta)}
}
\value{
PCA results object with eigenvalue plot

A list containing the PCA model. Side effect of generating a scree
plot.
}
\description{
Methods from Boone and Staver 2020 Chapter 2: eigenvalue less than 2 means unidimensional.
Conducts a PCA of standardized residual correlations.

Test construct unidimensionality using Principal Components
Analysis R(?) (PCAR).
}
\examples{
\dontrun{
# Test unidimensionality
uni_results <- test_uni_pcar(rasch_model)
}
\dontrun{
  model <- eRm::RM(df)
  test_uni_pcar(
    model = model,
    rotate = 'Promax',
    n_factors = 7
  )
}
}
\seealso{
\code{\link[=get_fits]{get_fits()}}, \code{\link[=test_lr_gender]{test_lr_gender()}}, \code{\link[=test_lr_rurality]{test_lr_rurality()}},
\code{\link[=test_lr_race]{test_lr_race()}}
}
