% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fits.R
\name{get_fits}
\alias{get_fits}
\title{Get person and item fits}
\usage{
get_fits(model)
}
\arguments{
\item{model}{An eRm model object.}
}
\value{
A list containing person fits and item fits.
}
\description{
Using an eRm model, get person parameter fit and item fit
objects.
}
\examples{
\dontrun{
  model <- eRm::RM(df)
  fits <- get_fits(model)
}
}
\seealso{
\code{\link[=test_lr_gender]{test_lr_gender()}}, \code{\link[=test_lr_rurality]{test_lr_rurality()}}, \code{\link[=test_lr_race]{test_lr_race()}}
}
