% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasch_model.R
\name{get_rasch_model}
\alias{get_rasch_model}
\title{Get Rasch Model}
\usage{
get_rasch_model(
  df,
  id,
  rebl_items,
  type = c("cml", "mml_constrained", "mml_con", "mml_unconstrained", "mml_uncon", "2pl",
    "tpm")
)
}
\arguments{
\item{df}{A dataframe containing the REBL items and participant IDs}

\item{id}{A character string specifying the column name containing
participant IDs}

\item{rebl_items}{A character vector of REBL item column names to include in
the model. Get this with \code{id_rebl_items()} if you don't have it handy.}

\item{type}{A string. \code{cml} = conditional maximum likelihood, \code{mml_con} =
constrained maximum marginal likelihood, \code{mml_uncon} = unconstrained
maximum marginal likelihoood, \code{tpl} = two parameter logistic model, \code{tpm} =
three parameter logistic model. See details for more info.}
}
\value{
An eRm Rasch model object that can be passed to other functions like
\code{\link[=get_rebl_scores]{get_rebl_scores()}}
}
\description{
Fit a Rasch model to REBL items using the eRm package. This
creates a Rasch model object that can be used to calculate person
parameters (REBL scores) and item parameters.
}
\details{
Add some more details about what this means.
}
\examples{
\dontrun{
  # Get REBL items
  rebl_items <- id_rebl_items(raw_example, '^(?!res).*')

  # Fit Rasch model
  model <- get_rasch_model(raw_example, "id", rebl_items)
}
}
\seealso{
\code{\link[=id_rebl_items]{id_rebl_items()}}, \code{\link[=id_reversed_rebl_items]{id_reversed_rebl_items()}},
\code{\link[=reverse_code_rebl_items]{reverse_code_rebl_items()}}, \code{\link[=get_rebl_scores]{get_rebl_scores()}}
}
