% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_rebl.R
\name{recode_rebl}
\alias{recode_rebl}
\title{Recode REBL Items}
\usage{
recode_rebl(df)
}
\arguments{
\item{df}{Dataframe with cases excluded}
}
\value{
Same dataframe with "Yes" and "No" coded to 1 and 0 as factors based
on which one is the PEB
}
\description{
Recoding values of REBL items based on which option is the Pro-Environmental Behavior (PEB).
This applies to surveys 2a and 2b where responses need to be converted from "Yes"/"No" to
1/0 based on whether the response represents pro-environmental behavior.
}
\details{
Will intentionally throw an error if any of the required columns don't exist.
The function loads REBL item coding information from '5_objects/cleaning/rebl_item_coding.rds'
to determine which items should be standard coded vs reverse coded.
}
\examples{
\dontrun{
# Example usage
recoded_data <- recode_rebl(raw_survey_data)
}
}
