% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_uni_pcar.R
\name{test_uni_pcar}
\alias{test_uni_pcar}
\title{Test unidimensionality with PCAR method}
\usage{
test_uni_pcar(model, rotate = "Promax", n_factors = 5)
}
\arguments{
\item{model}{An eRm model object.}

\item{rotate}{Name of rotation to use for PCAR. Passed to \code{pca()}.}

\item{n_factors}{Number of factors to extract}
}
\value{
A list containing the PCA model. Side effect of generating a scree
plot.
}
\description{
Test construct unidimensionality using Principal Components
Analysis R(?) (PCAR).
}
\examples{
\dontrun{
  model <- eRm::RM(df)
  test_uni_pcar(
    model = model,
    rotate = 'Promax',
    n_factors = 7
  )
}
}
\seealso{
\code{\link[=get_fits]{get_fits()}}, \code{\link[=test_lr_gender]{test_lr_gender()}}, \code{\link[=test_lr_rurality]{test_lr_rurality()}},
\code{\link[=test_lr_race]{test_lr_race()}}
}
